+++ ChangeLog.md
@@ -40,6 +40,9 @@ This project's release branch is `master`. This log is written from the perspect
   * Move `LargeObjectId` to `psql-simple-class`.
   * Remove the `Rhyolite.TH` module. Use `file-embed` instead.
   * Move `Rhyolite.Backend.EmailWorker` to `Rhyolite.DB.Groundhog.EmailWorker` in `groundhog-legacy`.
+  * Move the `Account` type to `groundhog-legacy-types` and move associated backend functionality to `Rhyolite.DB.Groundhog.Account` in `groundhog-legacy`.
+  * Move `Signed` and `MonadSign` to `signed-data`. Move `signWithKey`, `readSignedWithKey`, `SignT`, etc to `signed-data-clientsession`. Move groundhog instances for `MonadSign` and `SignT` to `Rhyolite.DB.Groundhog.Orphans`.
+  * Move `Account` to `groundhog-legacy-types`. Move backend code for accounts, e.g., `ensureAccountExists` to `Rhyolite.Account.Groundhog` in `groundhog-legacy`.
 * New:
   * Add a `Psql` instance for beam's `Pg`
 * Version bumps: